+++ Extensions/IISWebAppDeploy/Src/Tasks/IISWebAppDeploy/Main.ps1
@@ -17,6 +17,7 @@ param (
 
 $env:CURRENT_TASK_ROOTDIR = Split-Path -Parent $MyInvocation.MyCommand.Path
 
+. $env:CURRENT_TASK_ROOTDIR\TelemetryHelper\TelemetryHelper.ps1
 . $env:CURRENT_TASK_ROOTDIR\DeployIISWebApp.ps1
 
 (Main -machinesList $machinesList -adminUserName $adminUserName -adminPassword $adminPassword -winrmProtocol $winrmProtocol -testCertificate $testCertificate -webDeployPackage "$webDeployPackage" -webDeployParamFile "$webDeployParamFile" -overRideParams "$overRideParams" -websiteName "$websiteName" -removeAdditionalFiles $removeAdditionalFiles -excludeFilesFromAppData $excludeFilesFromAppData -takeAppOffline $takeAppOffline -additionalArguments $additionalArguments -deployInParallel $deployInParallel)